ref(normalization): Move exception to normalization
