+++ relay-event-normalization/src/event.rs
@@ -7,6 +7,8 @@ use std::collections::hash_map::DefaultHasher;
 use std::hash::{Hash, Hasher};
 use std::mem;
 
+use once_cell::sync::OnceCell;
+use regex::Regex;
 use relay_base_schema::metrics::{
     can_be_valid_metric_name, DurationUnit, FractionUnit, MetricUnit,
 };
@@ -568,6 +570,7 @@ fn normalize_exceptions(event: &mut Event) {
             //
             // We also want to validate some other aspects of it.
             for exception in exceptions {
+                normalize_exception(exception);
                 if let Some(exception) = exception.value_mut() {
                     if let Some(mechanism) = exception.mechanism.value_mut() {
                         mechanism::normalize_mechanism(mechanism, os_hint);
@@ -576,6 +579,35 @@ fn normalize_exceptions(event: &mut Event) {
             }
         }
     }
+}
+
+fn normalize_exception(exception: &mut Annotated<Exception>) {
+    static TYPE_VALUE_RE: OnceCell<Regex> = OnceCell::new();
+    let regex = TYPE_VALUE_RE.get_or_init(|| Regex::new(r"^(\w+):(.*)$").unwrap());
+
+    let _ = processor::apply(exception, |exception, meta| {
+        if exception.ty.value().is_empty() {
+            if let Some(value_str) = exception.value.value_mut() {
+                let new_values = regex
+                    .captures(value_str)
+                    .map(|cap| (cap[1].to_string(), cap[2].trim().to_string().into()));
+
+                if let Some((new_type, new_value)) = new_values {
+                    exception.ty.set_value(Some(new_type));
+                    *value_str = new_value;
+                }
+            }
+        }
+
+        if exception.ty.value().is_empty() && exception.value.value().is_empty() {
+            meta.add_error(Error::with(ErrorKind::MissingAttribute, |error| {
+                error.insert("attribute", "type or value");
+            }));
+            return Err(ProcessingAction::DeleteValueSoft);
+        }
+
+        Ok(())
+    });
 }
 
 fn normalize_user_agent(_event: &mut Event, normalize_user_agent: Option<bool>) {
@@ -984,6 +1016,7 @@ mod tests {
     use std::collections::BTreeMap;
 
     use insta::assert_debug_snapshot;
+    use itertools::Itertools;
     use relay_event_schema::protocol::{
         Contexts, Csp, DeviceContext, Event, Headers, IpAddr, Measurements, Request, Tags,
     };
@@ -2514,5 +2547,76 @@ mod tests {
         });
 
         assert_eq!(user.other, Object::new());
+    }
+
+    #[test]
+    fn test_handle_types_in_spaced_exception_values() {
+        let mut exception = Annotated::new(Exception {
+            value: Annotated::new("ValueError: unauthorized".to_string().into()),
+            ..Exception::default()
+        });
+        normalize_exception(&mut exception);
+
+        let exception = exception.value().unwrap();
+        assert_eq!(exception.value.as_str(), Some("unauthorized"));
+        assert_eq!(exception.ty.as_str(), Some("ValueError"));
+    }
+
+    #[test]
+    fn test_handle_types_in_non_spaced_excepton_values() {
+        let mut exception = Annotated::new(Exception {
+            value: Annotated::new("ValueError:unauthorized".to_string().into()),
+            ..Exception::default()
+        });
+        normalize_exception(&mut exception);
+
+        let exception = exception.value().unwrap();
+        assert_eq!(exception.value.as_str(), Some("unauthorized"));
+        assert_eq!(exception.ty.as_str(), Some("ValueError"));
+    }
+
+    #[test]
+    fn test_rejects_empty_exception_fields() {
+        let mut exception = Annotated::new(Exception {
+            value: Annotated::new("".to_string().into()),
+            ty: Annotated::new("".to_string()),
+            ..Default::default()
+        });
+
+        normalize_exception(&mut exception);
+
+        assert!(exception.value().is_none());
+        assert!(exception.meta().has_errors());
+    }
+
+    #[test]
+    fn test_json_value() {
+        let mut exception = Annotated::new(Exception {
+            value: Annotated::new(r#"{"unauthorized":true}"#.to_string().into()),
+            ..Exception::default()
+        });
+
+        normalize_exception(&mut exception);
+
+        let exception = exception.value().unwrap();
+
+        // Don't split a json-serialized value on the colon
+        assert_eq!(exception.value.as_str(), Some(r#"{"unauthorized":true}"#));
+        assert_eq!(exception.ty.value(), None);
+    }
+
+    #[test]
+    fn test_exception_invalid() {
+        let mut exception = Annotated::new(Exception::default());
+
+        normalize_exception(&mut exception);
+
+        let expected = Error::with(ErrorKind::MissingAttribute, |error| {
+            error.insert("attribute", "type or value");
+        });
+        assert_eq!(
+            exception.meta().iter_errors().collect_tuple(),
+            Some((&expected,))
+        );
     }
 }
